# openapi/media-svc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/media-items": {
      "post": {
        "operationId": "MediaItemController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItem"
                }
              }
            }
          }
        },
        "tags": ["media-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "MediaItemController_findAll",
        "parameters": [
          {
            "name": "text",
            "required": false,
            "in": "query",
            "allowEmptyValue": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "explode": true,
            "allowEmptyValue": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["media-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/media-items/{mediaId}": {
      "put": {
        "operationId": "MediaItemController_update",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMediaItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItem"
                }
              }
            }
          }
        },
        "tags": ["media-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "MediaItemController_remove",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["media-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "MediaItemController_findOne",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItemDto"
                }
              }
            }
          }
        },
        "tags": ["media-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/media-items/{mediaId}/share/{userId}": {
      "post": {
        "operationId": "MediaItemController_share",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": ":userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareItem"
                }
              }
            }
          }
        },
        "tags": ["media-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/media-items/popular": {
      "get": {
        "operationId": "MediaItemController_findPopular",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["media-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Media Service",
    "description": "Media Service",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "local"
    },
    {
      "url": "https://mediashare-api-staging.herokuapp.com",
      "description": "staging"
    },
    {
      "url": "https://mediashare-api-prod.herokuapp.com",
      "description": "production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "TagKeyValue": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": ["key", "value"]
      },
      "CreateMediaItemDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "UV3v8LhYtuA0CG7MBDdTi8m05zUrkzZV11m7lCZdaq3yHzFStMxppxhGka02Wr5vobXOEEScBphoJEyokvuCL8rHmZBQsctCWm7Zy1QXYDcPhotWlnpPn4xMRbS8l",
            "maxLength": 255,
            "minLength": 0
          },
          "userId": {
            "type": "string",
            "example": "cRbrYtxrLIeWoYhrz6WT8uU9ufMeRuo9vQ0tVLQvDKPDxnBKiN17HFtHrE9ndEDTYMLiIks1ir9YZ0UjKtCwRtaeLR2EeT12Cp6Div95JlNAWIqi0dDIV6HFSS6FJ",
            "maxLength": 255,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "example": "PPwNIqNrLH0u6DXDTXf7YqXgmLAu69vbwHcDhwptZnzdIQAlQ7yZtQ4g9uWQdgxCkKa6dz2wTdlzUeMv7XkJCGwPhHKNgHiCT3uqE7TjJGv52d17D9cqOO4mDeWGC",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "EETVJJBVdeaQGrYXPnfWv9Lo2s6CTx2PD3lBL1elAdUA5YREuWt0Ufyey09WhWcpKeUKiwIksd2187reUP0ap7BBHXztFBz4uFl1aHhmfT0KNGX4Y4qqwGFTyZqipJYk6t2L5LUH51bdzRmauHdiSUIkdpW0hDa5bDjHHb62JY1wQhkrMnwJMScZ1a5Cadwpe4hSQbbZZFcooayxuwag7bE4Hk4Yq2Qt1hOoYlqwgZCSO1v0iQwBK6Uh5MY5T9bF9XTSy945WerfPte4y5rHcsVNhvwNK10w1OWw0DUMIKev"
          },
          "description": {
            "type": "string",
            "example": "tawGYyYb96oFI6aMa6ryzqpywGMgfPcfUgauuv1y4dkfIgsT5wLxIk5GlNzcAISbYUPYNbfH6N0SApXipQO1xydoR3FM3MOHmjLF0h9V1r1XR2HqL3Fv3wVppCRVsSrBavqQttZj90waOHrDeX9j1zxPbBkcmk3MSF4P8VW0VHmevqa9aNvLuyLX6mXigbulnh7aQjZUhmigpHhyvggrE5y7eF2Z3RNwEnf3cXOGzEY5ATLvsD5FJfHrk9gDCcRi18b98BsX2Ray9q3hKLex9SmviJXt7ObDzVcuCpp7mrko"
          },
          "uri": {
            "type": "string",
            "example": "5F6lMNFLnr99dKJbPpHFU994xs63MH6kg3B2ofnr3xDbtuz3046IVpNDYrD909i6Yv6EfK5Tmrz6YYr49JiQcVRwk8Kvx8pdr1p099gDHCF8rXlr03Gk0WIbyCHAdbXOccwIfIBYIEwiOLQfLuziJFHRUF1hIbVAn8V7HGm6ug6Po9fGontjeYSmrmqcJP0pS5nuncGlrPKkimRYkQXsN3YOtDuEw2aut1yd5SJxQTLWUgGhKl77WNWZnO9WDSaNyPVahyf1syqfNpGPOQkTsdDfFGh4gQLfGprQod1Vw2D4"
          },
          "imageSrc": {
            "type": "string",
            "example": "JkNKhWpl0GPUf7tK5qYenbmmEqTQpG9bVlj8Kz9yfXv5HgZMT3Lht9laPSTYLHTU7M7voq4X5Zxxls9nMpMP2QRZTRi6Q2o0qXoYQfrr494eNY0Xzw9skkVqK271VutPiF6ebKaKcQ9vxOxtW1zetr5EyUF6Eerq86zAgWE2cgX0j0lcjqSGWiqXR5njSBbZ903LPyiHtnnRs18HH7AV0Gromj5KwzOxfH4J6tLlE0FdNnpcJ70G34nZAlqUDTGd2BFqrxn5YtnFe12l1FNMImhA6Du4XBscaF9FwCC5KIDH"
          },
          "isPlayable": {
            "type": "boolean"
          },
          "visibility": {
            "type": "string",
            "enum": ["private", "shared", "subscription", "public"]
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "eTag": {
            "type": "string",
            "example": "kgFkpVxM80ZJZpZZpEE62LoDfkwZEHzNsbPqy6kqISograhSzVbiGlcTGxUbxkRBiINHwFgZCTkFX1U9myava46UhoTR42gyQS8sAmmBGPrQwQJflelWtvgFpiTSB",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "key",
          "userId",
          "title",
          "description",
          "uri",
          "visibility"
        ]
      },
      "MediaItem": {
        "type": "object",
        "properties": {}
      },
      "MediaVisibilityType": {
        "type": "string",
        "enum": ["private", "shared", "subscription", "public"]
      },
      "UpdateMediaItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "JWuZMCJwXFLK53lVTQ8CJiC7IWxujnylov4h9AwYvu2ar8O2a0TlJiERkTMKGcx3q4hsBsaoqVHMmrFLs25NmF7YGG51ZOZ9VTTeeMJsJkPxVvFH03CMohFcaz2bp",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "title": {
            "type": "string",
            "example": "bfsf0YgVSS3jogV5Xcn0Kqpw8ydUk9LF6cWuSELOHBjt6gmHltDBYE1L7xX3CoWW36qCTy6FKlTv6V1Zn8ZaFLxKnTlYi0CuNoXCyY74WWQbw7Ufr4XK9OLzNhWn6",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "NYQmTAcPOfz0dPbPARpA7Q4JgGMypQRb0ybwZSexTHBt9GqRJ2WGkh9ca8czIIHEoASerfMPL1C9uOXPmTSHZfU7qq2WyBmLbZmxaAveOCc3daOC6s3OhZlhkDD4NThMZIKdoBjB5AMN8JIMTgN3rGQeprKJh0SJHoz057VozEKQItG42nQGkd0uMfY4MfNijIiF99fojGQFPHnYKKFUgxFwN6mFQ0xYVQAxzem3KQTssiwphe0anH1vntRuFI0GVP2vNOteyfzwWS5o4IZ0mVzMMqU9E9gfIDOQt2ybEago"
          },
          "description": {
            "type": "string",
            "example": "AKF1DJa84wnTqPpj4bAyHW6kIGZF8FsZlvUctpSV0imPM1OS4yw2exXtxSrBMdwMhDExgfYMRe3O3v5093f9tGQUHMqGtHLhv4JqMvIsLVEOZ5fQ6Gqso8dtK86byvjtXgqE6S9bEyVVmrNcVBifdXD37rv7RkiDLHmduPyle6CKpc5Gr6muFR051BJtyp4QQSFt26hUNHZ6aDwt6W0A8dFJiZcwM8g5aw62H5M1cgseD1EF1kd2DDgz4mPQGT5jAADsKjEZcLFxPX29bqKTdfEnvb7WiPW4DTwmOdlSR6dn"
          },
          "imageSrc": {
            "type": "string",
            "example": "5wfjXzEIcnkKXza2uCSm8Rzce90CDRzDi5PzicQYqsvRJGAtChQfcxhiOkcAwZcUSdz2yv6986VFIT9G1gq0cQheyjwwBdbZgpZZMuXLQUW10kBnHl8hovdPI1Fx7HjvEAa2jSVQccAAQHLxD0jkSSHZ13Y7tgIVlW3kpn9SMJIYzF7lBk054aODsUk9drjV9IqrCcSALyqM5tABPoCK0TqSllC2lSSH65or9ezOgYxyJXQdZwC7FnCDHRMFJGe38CVODiNJHAyMNaLfuTfFes7V2n5pF47TlTrwo6ck4Kdl"
          },
          "visibility": {
            "$ref": "#/components/schemas/MediaVisibilityType"
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "eTag": {
            "type": "string",
            "example": "rUBDmBaQlBqSsKyzD8cbqeRch8JPrUDGYmAdHm50bEdKMYKQvSI1edlGk2mnS0TsQy5pUQiqJ1eVuxVN5na9y8qiB9sjnqRgBuLhc79MFdnL3rhdPAKCiT3BSfqS1",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": ["_id", "title", "description", "visibility"]
      },
      "ShareItem": {
        "type": "object",
        "properties": {}
      },
      "MediaItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "JWuZMCJwXFLK53lVTQ8CJiC7IWxujnylov4h9AwYvu2ar8O2a0TlJiERkTMKGcx3q4hsBsaoqVHMmrFLs25NmF7YGG51ZOZ9VTTeeMJsJkPxVvFH03CMohFcaz2bp",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "key": {
            "type": "string",
            "example": "PEnT29xwgPHwqcinagxjtSebGHJFv8fF7xpmTmt9d9T2wxyXatOYhzedLZsPRCWeFnVNiqdIQG86R6AoMr1PbShgC4j1qTYqvitek6ivy3fe9wTPqUzKVOHjveeVw",
            "maxLength": 255,
            "minLength": 0
          },
          "userId": {
            "type": "string",
            "example": "25fIJNU6L1BQ6NgHm5phRCi1rnZxwMdnZryAyp3DXmw2qBQWSFw6axXx9cdJ7DYPPJvkAgQIBTshQ2qtEqkNYa6vQL8v1xwSUMvJl8qmo25MXoDG0aZBPauZpdChw",
            "maxLength": 255,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "example": "eNfFpVtqu7rVNRxHkbQoMBofYsYkh26gDtrg7Q1GWNGcfWsx5cItamMONyVvLt0PwuTgFJXqKcMEhzAMRErxd8FY0mZoaqRvoZfVBfa6S7Ow9JHAxGKS807etOY9u",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "U3mzMkkeFpp7MPAGshiZd3vQPqBrizidERvhAEEYqJziWtK4Odz9TWFUcjHxCDzB2AhYN0tsaYnwoRQH60FmHVJ5RS5XpmuX7Eb9c9qYPMrMbkmcjVg6ko3ekrmpfCBku40zKjSEPjc3CMqz3fqgWPwYXTtp7QWwRbO0AnScoowZuAIpCc1rx9EOYp8SU29AWiTBn8QXWd9NKZOCKO81Implx4SaNU1go7RnDWLGMpTNWLVvTcbpWtFrt1zqTDGnEuQZUCnUwxvDnXxTlf2zyK14m7kqTS6QzsqX262ibVZi"
          },
          "description": {
            "type": "string",
            "example": "azUbo2T9JfKhFbR3kXOeNyE7xhaAlXLV5WUlax3jH6NxgVZv0GR65WnWhubmqSa9PtIoEiE3BCdittGlygqPbeA6pCMr3KGCXOsYuVdVY2mzjku09MlKj7nQHEESAVTrHHZNtyEUSft8VnWm9MkbLoQKD7svAA4i5yFktkkjguYIo1VbWZQp8bxClpwGS43clXCrQY7FUXwKAUcy5H9wQzULMu1q5c8rQnf4KctF8nM20dznqJWVdDJQ4jMtm7f0dBltGxWBDD1dP66tV8JYTOJL0paVNNiO95AbUWC6FYxY"
          },
          "uri": {
            "type": "string",
            "example": "7HuRvCG53qk2PWjS0FHvlpEJqFVxfbKlNvbb3NvfkNiPfvSMAOHcx2PkXXr7IPPRyMByYSN0bRXXNJBl8B2pbWa4PKu4iVuCJmsvwIB1JcBcfkvrrlGV5Q9QdFosYvfM9MuA8UMNsZwyVQ6NKKyPP3oJ7oH2iIhK6ekOKA9g019OYY9R0wytw9mQpC8rhIxTyA0FGxOPjrHGiGFpEODjavqhLPegCEtXu7aMo2Np1hjRBo1nW9qsrewox0WtaB4KnmqpHqzTd12qP30lW56jA4YD0EkpgOl3Fx24rOtc4DL5"
          },
          "imageSrc": {
            "type": "string",
            "example": "4TSbWybuQ3lqQ9TIYNvUOuNPdxtN5CnQURhHojOFCOMQAmbhqGPF4Ea4xqsxo0j5TbJOPUCi4mq3tmWytMXYdugnrgFatGxpuf0AcJI11IzvVGwUBToBZefR8Tm3Ic2R9xUXUzPfj45WAW3SS95Ft3g1MCl3CHraJ8TczbJJcjHIYFoBsxnD98I441Gabvd17oXmYs7tFfXHD0QxQNDS3EGRsAuEio1FQONCZxgoAXtvfdk3uZXxhMoMUfdYYhPNAB64dvFOQZDcYx0kvRGRvhodswEJBGKLoeFcn8VgwvmN"
          },
          "isPlayable": {
            "type": "boolean"
          },
          "visibility": {
            "type": "string",
            "enum": ["private", "shared", "subscription", "public"]
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "eTag": {
            "type": "string",
            "example": "4iQkgfeFMUNDEYkRZHNhNo22jrrEM9mrf9XcVl5Xi0Funw34dMmYgF10Iu9K9dSNQXmPjWM7sMiA5yx480SNZGRx8kCOmjlbAEsR9M9MKn8pYBup91MdbH7bjdEXY",
            "maxLength": 255,
            "minLength": 0
          },
          "shareCount": {
            "type": "number"
          },
          "viewCount": {
            "type": "number"
          },
          "likesCount": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "key",
          "userId",
          "title",
          "description",
          "uri",
          "visibility",
          "shareCount",
          "viewCount",
          "likesCount"
        ]
      },
...
    }
  }
}